services/ai_server/llm/requirements.txt
flask>=3.0.0
mysql-connector-python>=8.0.0
